Speedup src/dbnode/client/ tests
